Review update: Extract attribute UTs to different file with no changes
